requirements.txt
# ============================================================
# RLRF: Reinforcement Learning from Rendering Feedback
# Requirements — tested on Python 3.10, CUDA 11.8+
# For Kaggle T4: install with pip install -r requirements.txt
# ============================================================

# Core ML
torch>=2.2.0
torchvision>=0.17.0

# HuggingFace ecosystem
transformers>=4.49.0
accelerate>=0.30.0
datasets>=2.18.0
peft>=0.10.0           # LoRA / QLoRA adapters
trl>=0.8.6             # GRPOTrainer, SFTTrainer

# Quantization (4-bit on Kaggle T4)
bitsandbytes>=0.43.0

# SVG rendering (paper uses CairoSVG)
cairosvg>=2.7.1

# Image processing & rewards
Pillow>=10.0.0
opencv-python-headless>=4.9.0  # headless for Kaggle/server
numpy>=1.26.0
scikit-image>=0.22.0   # SSIM metric

# Perceptual metrics
lpips>=0.1.4           # LPIPS perceptual loss (evaluation)

# Semantic reward: DreamSim (optional, ~1 GB checkpoint)
# dreamsim>=0.1.0      # Uncomment if VRAM allows; else SSIM fallback used

# CLIP (for Text2SVG and fallback semantic reward)
open-clip-torch>=2.24.0

# Utilities
scipy>=1.12.0
tqdm>=4.66.0
matplotlib>=3.8.0
pandas>=2.2.0
einops>=0.7.0

# Tokenizers (fast)
tokenizers>=0.19.0
sentencepiece>=0.2.0
